+++ src/utils/printers/types/PrintableMessageType.ts
@@ -1,0 +1,8 @@
+export type PrintableMessage = {
+    id: string;
+    bot: boolean;
+    tags?: string[];
+    card: boolean;
+    content: string;
+    created?: Date
+};